requirements.txt
aiogram==2.12.1
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
Babel==2.9.0
chardet==3.0.4
emoji==1.2.0
environs==8.0.0
gino==1.0.1
idna==2.10
psycopg2==2.8.6
python-dotenv==0.17.0
pytz==2020.5
SQLAlchemy==1.3.24
typing-extensions==3.7.4.3
urllib3==1.26.2
emoji-country-flag==1.2.4